1. refactor the UI screen.
2. merge similar ViewModels and Repositories.

3. set the file encoding of Android IDE as UTF-8 so that the Logcat can display Chinese text.

+++ app/src/main/java/com/example/healthhelper/dietary/util/graph/piechart/PieChartEntry.kt
@@ -1,9 +1,8 @@
 package com.example.healthhelper.dietary.util.graph.piechart
 
 import androidx.compose.ui.graphics.Color
-import java.math.BigDecimal
 
 data class PieChartEntry(
     val color: Color,
-    val percentage: BigDecimal,
+    val percentage: Double,
 )
